[QRF-24] Move `getUnreachableAncestors()` to the `slate-lists` package
It's never used anywhere else
