+++ packages/slate-commons/src/commands/index.ts
@@ -10,7 +10,6 @@ export { getNodePath } from './getNodePath';
 export { getNextSibling } from './getNextSibling';
 export { getPrevChars } from './getPrevChars';
 export { getPreviousSibling } from './getPreviousSibling';
-export { getUnreachableAncestors } from './getUnreachableAncestors';
 export { hasVoidElements } from './hasVoidElements';
 export { insertEmptyParagraph } from './insertEmptyParagraph';
 export { insertNodes } from './insertNodes';
