+++ packages/slate-lists/src/lib/increaseDepth.ts
@@ -5,6 +5,7 @@ import type { ListsEditor } from '../types';
 import { ListType } from '../types';
 
 import { getListItemsInRange } from './getListItemsInRange';
+import { getUnreachableAncestors } from './getUnreachableAncestors';
 import { increaseListItemDepth } from './increaseListItemDepth';
 import { wrapInList } from './wrapInList';
 
@@ -25,8 +26,8 @@ export function increaseDepth(editor: ListsEditor): void {
 
     // When calling `increaseListItemDepth` the paths and references to list items
     // can change, so we need a way of marking the list items scheduled for transformation.
-    const refs = EditorCommands.getUnreachableAncestors(indentableListItemsInRange).map(
-        ([_, path]) => Editor.pathRef(editor, path),
+    const refs = getUnreachableAncestors(indentableListItemsInRange).map(([_, path]) =>
+        Editor.pathRef(editor, path),
     );
 
     // Before we indent "list-items", we want to convert every non list-related block in selection to a "list".
